VZ-7295: moved more functions into status.go
