+++ platform-operator/controllers/verrazzano/status.go
@@ -9,14 +9,56 @@ import (
 	ctrlerrors "github.com/verrazzano/verrazzano/pkg/controller/errors"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	installv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/validators"
 	vzconst "github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/registry"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	vzcontext "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/context"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/vzinstance"
 	corev1 "k8s.io/api/core/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"time"
 )
+
+// checkInstallComplete checks to see if the install is complete
+func (r *Reconciler) checkInstallComplete(vzctx vzcontext.VerrazzanoContext) (bool, error) {
+	log := vzctx.Log
+	actualCR := vzctx.ActualCR
+	ready, err := r.checkComponentReadyState(vzctx)
+	if err != nil {
+		return false, err
+	}
+	if !ready {
+		return false, nil
+	}
+	// Set install complete IFF all subcomponent status' are "CompStateReady"
+	message := "Verrazzano install completed successfully"
+	// Status update must be performed on the actual CR read from K8S
+	return true, r.updateStatus(log, actualCR, message, installv1alpha1.CondInstallComplete)
+}
+
+// checkUpgradeComplete checks to see if the upgrade is complete
+func (r *Reconciler) checkUpgradeComplete(vzctx vzcontext.VerrazzanoContext) (bool, error) {
+	if vzctx.ActualCR == nil {
+		return false, nil
+	}
+	if vzctx.ActualCR.Status.State != installv1alpha1.VzStateUpgrading {
+		return true, nil
+	}
+	log := vzctx.Log
+	actualCR := vzctx.ActualCR
+	ready, err := r.checkComponentReadyState(vzctx)
+	if err != nil {
+		return false, err
+	}
+	if !ready {
+		return false, nil
+	}
+	// Set upgrade complete IFF all subcomponent status' are "CompStateReady"
+	message := "Verrazzano upgrade completed successfully"
+	// Status and State update must be performed on the actual CR read from K8S
+	return true, r.updateVzStatusAndState(log, actualCR, message, installv1alpha1.CondUpgradeComplete, installv1alpha1.VzStateReady)
+}
 
 // updateStatus updates the status in the Verrazzano CR
 func (r *Reconciler) updateStatus(log vzlog.VerrazzanoLogger, cr *installv1alpha1.Verrazzano, message string, conditionType installv1alpha1.ConditionType) error {
@@ -126,6 +168,65 @@ func (r *Reconciler) updateComponentStatus(compContext spi.ComponentContext, mes
 	return r.updateVerrazzanoStatus(log, cr)
 }
 
+// Convert a condition to a VZ State
+func conditionToVzState(currentCondition installv1alpha1.ConditionType) installv1alpha1.VzStateType {
+	switch currentCondition {
+	case installv1alpha1.CondInstallStarted:
+		return installv1alpha1.VzStateReconciling
+	case installv1alpha1.CondUninstallStarted:
+		return installv1alpha1.VzStateUninstalling
+	case installv1alpha1.CondUpgradeStarted:
+		return installv1alpha1.VzStateUpgrading
+	case installv1alpha1.CondUpgradePaused:
+		return installv1alpha1.VzStatePaused
+	case installv1alpha1.CondUninstallComplete:
+		return installv1alpha1.VzStateReady
+	case installv1alpha1.CondInstallFailed, installv1alpha1.CondUpgradeFailed, installv1alpha1.CondUninstallFailed:
+		return installv1alpha1.VzStateFailed
+	}
+	// Return ready for installv1alpha1.CondInstallComplete, installv1alpha1.CondUpgradeComplete
+	return installv1alpha1.VzStateReady
+}
+
+// setInstallStartedCondition
+func (r *Reconciler) setInstallingState(log vzlog.VerrazzanoLogger, vz *installv1alpha1.Verrazzano) error {
+	// Set the version in the status.  This will be updated when the starting install condition is updated.
+	bomSemVer, err := validators.GetCurrentBomVersion()
+	if err != nil {
+		return err
+	}
+
+	vz.Status.Version = bomSemVer.ToString()
+	return r.updateStatus(log, vz, "Verrazzano install in progress", installv1alpha1.CondInstallStarted)
+}
+
+// checkComponentReadyState returns true if all component-level status' are "CompStateReady" for enabled components
+func (r *Reconciler) checkComponentReadyState(vzctx vzcontext.VerrazzanoContext) (bool, error) {
+	cr := vzctx.ActualCR
+	if unitTesting {
+		for _, compStatus := range cr.Status.Components {
+			if compStatus.State != installv1alpha1.CompStateDisabled && compStatus.State != installv1alpha1.CompStateReady {
+				return false, nil
+			}
+		}
+		return true, nil
+	}
+
+	// Return false if any enabled component is not ready
+	for _, comp := range registry.GetComponents() {
+		spiCtx, err := spi.NewContext(vzctx.Log, r.Client, vzctx.ActualCR, nil, r.DryRun)
+		if err != nil {
+			spiCtx.Log().Errorf("Failed to create component context: %v", err)
+			return false, err
+		}
+		if comp.IsEnabled(spiCtx.EffectiveCR()) && cr.Status.Components[comp.Name()].State != installv1alpha1.CompStateReady {
+			spiCtx.Log().Progressf("Waiting for component %s to be ready", comp.Name())
+			return false, nil
+		}
+	}
+	return true, nil
+}
+
 // initializeComponentStatus Initialize the component status field with the known set that indicate they support the
 // operator-based install.  This is so that we know ahead of time exactly how many components we expect to install
 // via the operator, and when we're done installing.
@@ -177,6 +278,17 @@ func (r *Reconciler) initializeComponentStatus(log vzlog.VerrazzanoLogger, cr *i
 		return newRequeueWithDelay(), r.updateVerrazzanoStatus(log, cr)
 	}
 	return ctrl.Result{}, nil
+}
+
+// setUninstallCondition sets the Verrazzano resource condition in status for uninstall
+func (r *Reconciler) setUninstallCondition(log vzlog.VerrazzanoLogger, vz *installv1alpha1.Verrazzano, newCondition installv1alpha1.ConditionType, msg string) (err error) {
+	// Add the uninstall started condition if not already added
+	for _, condition := range vz.Status.Conditions {
+		if condition.Type == newCondition {
+			return nil
+		}
+	}
+	return r.updateStatus(log, vz, msg, newCondition)
 }
 
 func (r *Reconciler) updateVerrazzanoStatus(log vzlog.VerrazzanoLogger, vz *installv1alpha1.Verrazzano) error {
